LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-xy-in-scrolling-div-expected.txt

FAIL div.scrollTop should be 161. Was 0.
FAIL div.scrollLeft should be 322. Was 0.
